Final Collisions Changes
We could add more smoothing after.
