--- DuKoup/Assets/Scripts/Movement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5773c2138c296cc43969c29faadaa887
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
