+++ DuKoup/Assets/Scripts/PlayersCollision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c075a3495639a47f28c762eec59172ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
