+++ DuKoup/Library/StateCache/SceneView/c6090f-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":2.2014474868774416,"y":0.9721865653991699,"z":4.493744850158691},"rotation":{"x":-0.19345799088478089,"y":0.6014811396598816,"z":-0.15314114093780518,"w":-0.7598307132720947},"size":13.745861053466797,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":20.241073608398439,"y":-4.861431121826172,"z":26.240489959716798},"rotation":{"x":-0.09110037982463837,"y":-0.4227897524833679,"z":0.04276623576879501,"w":-0.9006239175796509},"size":74.34490203857422,"orthographic":false}
